- Add union message to protoc to enable simulation mode protocol - Add flow control for write to cassandra in patchdbmq server

+++ rhessys/main.cc
@@ -287,12 +287,21 @@
 /*--------------------------------------------------------------*/
 #include <signal.h>
 #include <stdio.h>
+#include <iostream>
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
 #include <unistd.h>
 
+#include <zmq.h>
+
 #include "rhessys.h"
+#include "functions.h"
+
+#include "zmqutil.h"
+#include "patch.pb.h"
+
+using namespace std;
 
 // The $$RHESSYS_VERSION$$ string will be replaced by the make
 // script to reflect the current RHESSys version.
@@ -505,6 +514,58 @@ int	main( int main_argc, char **main_argv)
 		printf("Stopping patchdb (%d) message queue server %s... ",
 				patchdb_pid,
 				command_line[0].patchdb_server);
+		//kill(patchdb_pid, SIGTERM);
+		rhessys::PatchDBMesg m;
+		m.set_type(m.END_SIM);
+		rhessys::EndSim *e = m.mutable_endsim();
+		e->set_mesg("E");
+
+		zmq_msg_t msg;
+		int rc = PbToZmq(&m, &msg);
+		if (rc == -1) {
+			printf("Unable to create zeromq message");
+			exit(EXIT_FAILURE);
+		}
+		// TODO: send as multipart mesg to we can send control
+		// information.
+		// See: http://zguide.zeromq.org/page:all#Multipart-Messages
+		rc = zmq_sendmsg(output->patchdbmq_requester,
+				&msg, 0);
+
+		if (rc == -1) {
+			if (errno == EFSM) {
+				printf("output_patch: zeromq: operation cannot be performed on this socket at the moment due to the socket not being in the appropriate state.\n");
+			} else {
+				printf("output_patch: zeromq returned, main:540: %s, exiting...\n", strerror(errno));
+			}
+			exit(EXIT_FAILURE);
+		}
+
+		//printf("Waiting for close response from zeromq server...");
+
+		char response[2];
+		rc = zmq_recv(output->patchdbmq_requester, response, 1, 0);
+		if (rc == -1) {
+			if (errno == EFSM) {
+				printf("output_patch: zeromq: operation cannot be performed on this socket at the moment due to the socket not being in the appropriate state.\n");
+			} else {
+				printf("output_patch: zeromq returned, main:550: %d, exiting...\n", errno);
+			}
+			exit(EXIT_FAILURE);
+		}
+
+		if (response[0] != 'A') {
+			printf("output_patch: expected patchdbmq server to return %s, but received: %s, exiting...\n",
+					"A", response);
+			exit(EXIT_FAILURE);
+		}
+
+		//printf("\nReceived %s from zeromq server\n", response);
+
+		// Close zeromq socket
+		zmq_close(output->patchdbmq_requester);
+		zmq_ctx_destroy(output->patchdbmq_context);
+
 		kill(patchdb_pid, SIGTERM);
 		printf("done.\n");
 	}
